### 2024-02-10 #6113 @dfm
@AndrewAnnex — thanks for the response! We aim for a turnaround of 4 weeks, but JOSS reviews are iterative so many take more or less time than that. We ask that reviewers start their review ASAP, and then revisit a few times throughout the revisions. Therefore a review doesn't typically take one large chunk of time. Hope this helps provide some context!
