### 2024-04-01 #6566 @editorialbot
[REVIEW]: OMEGA-Py: Python Tools for OMEGA Data
**Submitting author:** @AStcherbinine (<a href="http://orcid.org/0000-0002-7086-5443">Aurélien Stcherbinine</a>)
**Repository:** https://github.com/AStcherbinine/omegapy
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 3.1
