### 2024-04-03 AStcherbinine/omegapy#4 @Sierra-MC
No contribution guidelines  documentation
openjournals/joss-reviews#6566

I'm not seeing contribution guidelines documentation or information for users on how to report issues or seek support. For contribution guidelines you may wish to look at https://www.contributor-covenant.org/.

Language from JOSS review requirement: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
