### 2024-04-03 AStcherbinine/omegapy#5 @Sierra-MC
Missing example usage
Examples in the documentation are missing:

https://astcherbinine.github.io/omegapy/example1_band_depth/
https://astcherbinine.github.io/omegapy/example2_overplot/

I know there is the basic usage examples, but it would be really nice to see these examples as band depth especially is a very common operation that people will want to execute.

openjournals/joss-reviews#6566
