### 2024-04-03 AStcherbinine/omegapy#6 @Sierra-MC
Test coverage?
I'm not seeing test coverage in this repository. Is there an available test suite?

openjournals/joss-reviews/issues/6566
