### 2024-04-03 AStcherbinine/omegapy#6 @Sierra-MC
Referencing review requirement:
Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?
