### 2024-04-03 AStcherbinine/omegapy#5 @Sierra-MC
Referencing review requirement: Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
