### 2024-04-03 AStcherbinine/omegapy#7 @Sierra-MC
Add IDL Routines as a Citation instead of a hyperlink
Use a citation (which can include the url) instead of solely a url for the IDL Routines.

Review requirement: References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

openjournals/joss-reviews#6566
