### 2024-04-12 e77af04
Add community guidelines + update credits
