+++ docs/community.md
@@ -1,0 +1,37 @@
+## Contribute
+We are welcoming contributions to improve OMEGA-Py. 
+Please fork the [`omegapy`](https://github.com/AStcherbinine/omegapy) repository to make your 
+updates, then make a pull request to merge it with the main project branch.
+
+For each new update, please follow these guidelines:
+
+ * For any change to OMEGA-Py, update the documentation accordingly in *docs*.
+ * Keep the code style (the naming of functions/scripts, the headers, sections...) and comment
+    your updates.
+ * In the pull request, explains clearly what you changed/added and what it brings to the project.
+
+
+## Report issues
+Please report any bugs or issues you encounter with OMEGA-Py by creating a
+[New Issue](https://github.com/AStcherbinine/omegapy/issues) on GitHub.
+
+To help us address the issue, please include at minimum:
+
+ * Your operating system name and version.
+ * Any details about your local setup that might be helpful in troubleshooting.
+   (How do you use Python? With an IDE? In a Jupyter Notebook?)
+ * Expected *vs* obtained result.
+ * Detailed steps to reproduce the bugs.
+ * Include if applicable any output, error message or screenshot relevant to understand the issue.
+ * If requesting for a new feature, please describe it precisely and explain why it will be
+   useful.
+
+
+## Support
+If you need support using PyBWE, you can either:
+
+ * Look carefully at the [documentation](https://astcherbinine.github.io/omegapy/) 
+   (first step = *RTFM*).
+ * Open an [issue](https://github.com/AStcherbinine/omegapy/issues) on GitHub.
+ * Contact us directly at [aurelien@stcherbinine.net](mailto:aurelien@stcherbinine.net)
+
