+++ docs/credits.md
@@ -1,12 +1,15 @@
 # Credits
 
-© Aurélien Stcherbinine (2020–2023)
+© Aurélien Stcherbinine (2020–2024)
 
 Institut d'Astrophysique Spatiale (IAS), Université Paris-Saclay, CNRS, Orsay, France
 
 LATMOS/IPSL, UVSQ Université Paris-Saclay, Sorbonne Université, CNRS, Guyancourt, France
 
 Department of Astronomy and Planetary Science, Northern Arizona University, Flagstaff, AZ, USA
+
+Institut de Recherche en Astrophysique et Planétologie (IRAP), CNES, Université Toulouse III,
+CNRS, Toulouse, France
 
 
 ## License
