### 2024-04-12 AStcherbinine/omegapy#4 @AStcherbinine
Thank you for pointing that.
A *contributing guidelines* page has been added in the documentation at https://astcherbinine.github.io/omegapy/community/ along with a link in the `README` in commit e77af04c2c05cafb003fc58b12e8d68e20ceee5b
