### 2024-04-12 AStcherbinine/omegapy#4 @Sierra-MC
Small typo of old package name in community guidelines. I made a pull request to fix that. Otherwise this looks good to go.
