+++ joss-paper/paper.bib
@@ -60,3 +60,10 @@
   urldate = {2022-10-17}
 }
 
+@article{soft10,
+  title = {{SOFT} 10},
+  author = {{OMEGA} team},
+  year = {2004},
+  url =
+  {https://archives.esac.esa.int/psa/ftp/MARS-EXPRESS/OMEGA/MEX-M-OMEGA-2-EDR-FLIGHT-EXT8-V1.0/SOFTWARE/}
+}
