### 2024-04-12 15a124c
fix transcription of OMEGA team in reference
