+++ joss-paper/paper.bib
@@ -61,7 +61,7 @@
 }
 
 @article{soft10,
-  title = {{SOFT} 10},
+  title = {{SOFT} 10 {IDL} routines},
   author = {{{OMEGA} team}},
   year = {2004},
   url =
