### 2024-04-12 AStcherbinine/omegapy#7 @AStcherbinine
Thank you for pointing that.
A reference has been added in the paper instead of the sole link for the SOFT10 with commit 09c998d1fdddcc9fc9815c3f41a3958818f5952f

Updated by editorialbot:
👉📄 [Download article proof](https://raw.githubusercontent.com/openjournals/joss-papers/joss.06566/joss.06566/10.21105.joss.06566.pdf) 📄 [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.06566/joss.06566/10.21105.joss.06566.pdf) 📄 👈
