### 2024-04-15 a61d40d
fix typo in omega_mask arguments
