+++ docs/example1_band_depth.md
@@ -1,2 +1,161 @@
-**Coming soon.**
+Here is an example on how to compute and display a band depth on an OMEGA observation using
+OMEGA-Py.
+
+Let's assume you have downloaded the OMEGA data cube *ORB0979_2* from the PSA
+(or from these links: 
+[QUB](https://archives.esac.esa.int/psa/ftp/MARS-EXPRESS/OMEGA/MEX-M-OMEGA-2-EDR-FLIGHT-V1.0/DATA/ORB09/ORB0979_2.QUB),
+[NAV](https://archives.esac.esa.int/psa/ftp/MARS-EXPRESS/OMEGA/MEX-M-OMEGA-2-EDR-FLIGHT-V1.0/DATA/GEM09/ORB0979_2.NAV))
+and loaded the module and data as follows:
+~~~python
+import omegapy.omega_data as od
+import omegapy.omega_plots as op
+~~~
+
+## Step 1 - Loading the data
+Option 1: Load the binary files and apply the thermal and atmospheric corrections if needed.
+~~~python
+# Load the data cube
+omega0 = od.OMEGAdata('0979_2')
+# Apply thermal and atmospheric corrections
+omega = od.corr_therm_atm(omega0, npool=15)     # Adjust npool according to your system
+~~~
+
+Option 2: Load directly a previously saved [`OMEGAdata`](../reference/omega_data/#omega_data.OMEGAdata) object 
+(with or without the corrections already applied).
+~~~python
+# Load directly the already corrected data cube
+omega = od.autoload_omega('0979_2', therm_corr=True, atm_corr=True)
+~~~
+
+## Step 2 - Generating the data mask (optional)
+
+~~~python
+mask = od.omega_mask(
+    omega, 
+    hide_128=True, 
+    emer_lim=10, 
+    inci_lim=70, 
+    tempc_lim=-194, 
+    limsat_c=500
+    )
+~~~
+
+## Step 3 - Computing band depth at 1.5 μm
+
+Let's compute the water ice 1.5 μm band depth as defined in Poulet et al. (2007)[^1]:
+
+[^1]: F. Poulet, C. Gomez, J.-P. Bibring, et al. (2007).
+Martian surface mineralogy from Observatoire pour la Minéralogie, l'Eau, les Glaces et l'Activité on board the Mars Express
+spacecraft (OMEGA/MEx): Global mineral maps.
+*JGR, 112*, E08S02.
+[doi:10.1029/2006JE002840](https://doi.org/10.1029/2006JE002840)
+
+~~~python
+nbd_15 = BD_omega(omega, [1.5, 1.51], 1.30, 1.71, norm=True)
+~~~
+
+??? note "Source code for function `#!python BD_omega(omega, lam0, lamc1, lamc2, norm=True)`"
+    ~~~python
+    import omegapy.useful_functions as uf
+    import numpy as np
+
+    def BD_omega(omega, lam0, lamc1, lamc2, norm=True):
+        """Compute the band depth on an OMEGA observation cube.
+        Continuum linear between lamc1 and lamc2.
+
+        If an array is passed as argument for a wavelength value, the average is used.
+
+        Parameters
+        ==========
+        omega : OMEGAdata
+            The OMEGA/MEx observation.
+        lam0 : float or array-like
+            The wavelength of the center of the band.
+        lamc1 : float or array-like
+            The wavelength of the bluer point for the continuum determination.
+        lamc2 : float or array-like
+            The wavelength of the redder point for the continuum determination.
+        norm : bool, optional (default True)
+            | True -> band_depth output is the normalized BD values.
+            | False -> band_depth output is the BD values.
+
+        Returns
[…]
