### 2024-04-15 b6d6ac5
fix issue overlapping axes in show_data_v2 (similar to show_omega_v2)
