### 2024-04-16 25961e4
update doc alpha argument in omega_plots
