### 2024-04-16 AStcherbinine/omegapy#5 @AStcherbinine
Thank you for reminding me to finalize these examples. Both are now published online :

Band depth with commit 1876b099dc231bfd2d5375112140bfc984a9feb8 (https://astcherbinine.github.io/omegapy/example1_band_depth/)
And overplotting with commit 9065f76ba803699ad1897e822953508c330aaf7f (https://astcherbinine.github.io/omegapy/example2_overplot/)

Also, the source code of the `show_data_v2` function has been updated to fix a small bug when overplotting (new module version 3.0.1 published on PyPI)
