### 2024-04-16 92d4a38
move module version to JSON file
