+++ omegapy/__init__.py
@@ -16,3 +16,10 @@ useful_functions :
 name = 'omegapy'
 
 from . import omega_data, omega_plots, useful_functions
+
+import json
+
+with open('module_infos.json', 'r') as f:
+    data_json = json.load(f)
+
+__version__ = data_json['version']
