### 2024-04-16 9deb6c3
move Python version min to JSON file
