### 2024-04-16 0bd1c71
change Python version min entry name in JSON file
