### 2024-04-16 3d1bac8
change Python version min entry name in JSON file
