### 2024-04-16 7aca8a5
Update version badge - version from json file
