+++ README.md
@@ -1,4 +1,4 @@
-![version](https://img.shields.io/badge/version-3.0.1-blue)
+![version-json](https://img.shields.io/badge/dynamic/json?url=https%3A%2F%2Fraw.githubusercontent.com%2FAStcherbinine%2Fomegapy%2Fmaster%2Fpackage.json&query=%24.version&label=version&labelColor=grey&color=blue)
 ![pythonversion](https://img.shields.io/badge/Python-3.7+-blue)
 [![DOI](https://zenodo.org/badge/349763849.svg)](https://zenodo.org/doi/10.5281/zenodo.7818828)
 
