### 2024-04-16 199eefb
update Python-min badge in README from json file
