### 2024-04-16 #6566 @dfm
Thanks @ryanbanderson for agreeing to review! Please take a look at the info in the top comment of this thread, and [my first comment](https://github.com/openjournals/joss-reviews/issues/6566#issuecomment-2030218147) for more info about how the review will progress. As soon as possible please comment `@editorialbot generate my checklist` to get your review checklist and see if you can go through the first couple of items quickly. Thanks again!!
