### 2024-04-16 62c6a4e
update cube id in basic_usage
