### 2024-04-16 a7a92ae
add tests installation module in doc
