### 2024-04-16 AStcherbinine/omegapy#6 @AStcherbinine
Thanks for pointing that!

Commit a7a92aee203df2808344c3d099c756740b184a67 added a *Tests* in the documentation going over some tests to check the installation & configuration of the module, along with how to address some potential issues.
https://astcherbinine.github.io/omegapy/tests/
