### 2024-04-17 ad0e288
Add JOSS badge in README + update figure logo path to raw image
