### 2024-04-17 25495fd
add PSA reference + IDL SOFT10 functionality
