+++ joss-paper/paper.bib
@@ -62,8 +62,23 @@
 
 @article{soft10,
   title = {{SOFT} 10 {IDL} routines},
-  author = {{{OMEGA} team}},
+  author = {{{OMEGA} {Team}}},
   year = {2004},
   url =
   {https://archives.esac.esa.int/psa/ftp/MARS-EXPRESS/OMEGA/MEX-M-OMEGA-2-EDR-FLIGHT-EXT8-V1.0/SOFTWARE/}
 }
+
+@article{besse_2018,
+  title = {{{ESA}}'s {{Planetary Science Archive}}: {{Preserve}} and Present Reliable Scientific Data Sets},
+  shorttitle = {{{ESA}}'s {{Planetary Science Archive}}},
+  author = {Besse, S. and Vallat, C. and Barthelemy, M. and Coia, D. and Costa, M. and De Marchi, G. and Fraga, D. and Grotheer, E. and Heather, D. and Lim, T. and Martinez, S. and Arviset, C. and Barbarisi, I. and Docasal, R. and Macfarlane, A. and Rios, C. and Saiz, J. and Vallejo, F.},
+  year = {2018},
+  month = jan,
+  journal = {Planetary and Space Science},
+  series = {Enabling {{Open}} and {{Interoperable Access}} to {{Planetary Science}} and {{Heliophysics Databases}} and {{Tools}}},
+  volume = {150},
+  pages = {131--140},
+  issn = {0032-0633},
+  doi = {10.1016/j.pss.2017.07.013}
+}
+
