### 2024-04-17 #6566 @AStcherbinine
> @AStcherbinine,
>
> Would you be able to add a short summary in the paper of what the functionality of the original IDL script this is replacing does? Especially if this is targeting the younger generation of scientists who don't rely on IDL, there should be some explanation of what those routines did originally. While it may be obvious for users who used to use the IDL scripts and are moving to Python it is less so for new users of OMEGA data.

@Sierra-MC thank you for raising this point.
The second paragraph of the paper has been updated to include a brief statement of what the IDL SOFT10 routines are doing.
