### 2024-04-17 f9ddb7d
fix json path in __init__.py and bump to version 3.0.2
