### 2024-04-17 fcbe293
move package.json into omegapy rep and bump to version 3.0.3
