+++ setup.py
@@ -5,7 +5,7 @@ with open('README.md', 'r') as f:
     long_description = f.read()
 with open('requirements.txt', 'r') as f:
     requirements = f.read().strip('\n').split('\n')
-with open('package.json', 'r') as f:
+with open('omegapy/package.json', 'r') as f:
     data_json = json.load(f)
 
 package_data = {
