### 2024-04-25 AStcherbinine/omegapy#6 @Sierra-MC
I am finding an error with the version check on the tests page.
```
>>> import omegapy
>>> omegapy.__version__
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
AttributeError: module 'omegapy' has no attribute '__version__'
```
On the other hand if I get out of the python console and get back into the terminal I find:
```
(jossreview) sierra@pop-os:~/PycharmProjects/pdr/pdr$ pip freeze|grep omegapy
omegapy==3.0
```
as expected.
