### 2024-04-25 AStcherbinine/omegapy#6 @AStcherbinine
Yeah, I added this in a small patch. If you update to go to the current v3.0.4 it should work ;)
Just `pip3 install omegapy --upgrade`
