### 2024-04-25 298a1a2
fix path json raw file for badge in tests
