### 2024-04-25 AStcherbinine/omegapy#6 @Sierra-MC
```
# OMEGA file importation (ORB0979_3)
omega = od.OMEGAdata('0979_3')

# Atmospheric correction
omega_corr_atm = od.corr_atm(omega_corr)
```
Atmospheric correction in basic usage uses variable `omega_corr` which is not created beforehand. Might be that `omega` should be `omega_corr`. Or there is a missing step?
