### 2024-04-25 f886648
fix typo in basic usage
