+++ docs/basic_usage.md
@@ -10,14 +10,14 @@ import omegapy.useful_functions as uf
 omega = od.OMEGAdata('0979_3')
 
 # Atmospheric correction
-omega_corr_atm = od.corr_atm(omega_corr)
+omega_corr_atm = od.corr_atm(omega)
 
 # Simultaneous Atmospheric & Thermal corrections (for the use of the L-channel)
 # > Use the `npool` argument to control the number of simultaneous processes 
 # > used to compute the thermal correction 
 # > (e.g., npool=15 is usually a nice choice if your system can handle it)
 # > Note: multiprocessing is currently not available for Windows
-omega_corr_therm_atm = od.corr_therm_atm(omega_corr, npool=1)
+omega_corr_therm_atm = od.corr_therm_atm(omega, npool=1)
 
 # Thermal correction only
 omega_corr_therm = od.corr_therm(omega, npool=1)
