### 2024-04-25 AStcherbinine/omegapy#6 @AStcherbinine
Ooops, that's on me. It should be `omega` only for both
`omega_corr_atm = od.corr_atm(omega)` and `omega_corr_therm_atm = od.corr_therm_atm(omega, npool=1)` (some leftover from a previous organization of the example...)

Should be updated now.
