### 2024-04-25 AStcherbinine/omegapy#5 @Sierra-MC
Is the code for `BD_omega` in `omegapy` itself or is it only in the example usage as source code for users to recreate themselves?
