### 2024-04-25 AStcherbinine/omegapy#5 @AStcherbinine
Currently, it's only in the example. The idea was that the users can have the code to either use this function as it, or take it as an inspiration to implement their own band depth (or alternative spectral criteria) computation.
