### 2024-04-25 #6566 @Sierra-MC
@AStcherbinine, obviously you've  made major contributions to the software (you are the only contributor on github prior to this review). However, assessing the full list of authors is more difficult. Would you be able to provide a quick description of the contributions of the other authors for the following review criteria to be assessed?:

Has the submitting author (@AStcherbinine) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
