### 2024-04-25 #6566 @Sierra-MC
The tests here are a bit weak in that they mainly revolve around confirming installation and turn users to the basic usage rather than confirming the output of the functions being used are _accurate_. However they do _exist_ as a manual walk-through.
