### 2024-04-25 #6566 @AStcherbinine
> @AStcherbinine, obviously you've made major contributions to the software (you are the only contributor on github prior to this review). However, assessing the full list of authors is more difficult. Would you be able to provide a quick description of the contributions of the other authors for the following review criteria to be assessed?:
>
> Has the submitting author (@AStcherbinine) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

@Sierra-MC sure. So I am the only contributor for writing the Python code of the module but here are how these people contributed to this work:
* Y. Langevin: the one who wrote the original IDL software, so a large part of the module is from his work ported to Python. He also helped me understanding the original code and testing the accuracy of my Python implementation.
* J. Carter: current PI of the OMEGA instrument. Lead for making the official validation possible. Review the structure, attributes loaded in the `OMEGAdata` class and functionalities to make sure users will have everything they need.
* M. Vincendon: former PhD advisor, provide some additional IDL code to the sole SOFT10 that has been used in the module. Helped to go over the original IDL software and test the OMEGA-Py module
* Y. Leseigneur & O. Barraud: First two users of the module after me. By using it in their research, they provided a huge amount of constructive feedback, bug tracking, and suggestions that contributed to bring the module to its current state.
