### 2024-05-09 AStcherbinine/omegapy#9 @ryanbanderson
Recursive file search
Given the nested directory structure when data are downloaded from the PSA, it would be nice if calling
omega = od.OMEGAData('###_#') did a recursive search into subdirectories below the directory specified in omega_bin_path. Or at least add the option.
