### 2024-05-09 #6566 @ryanbanderson
@AStcherbinine I'm doing my review today, looks like a very useful tool! I just created an issue with a suggestion to enable a recursive option when searching for OMEGA data. Or maybe that option already exists and I just missed it...
https://github.com/AStcherbinine/omegapy/issues/9
