### 2024-05-09 #6566 @ryanbanderson
@AStcherbinine In the checklist it is asking whether there is a "statement of need" in the documentation. You have one in the paper, but it would be good to copy it over to the docs as well to expand on what is already there.
