### 2024-05-09 #6566 @ryanbanderson
> The tests here are a bit weak in that they mainly revolve around confirming installation and turn users to the basic usage rather than confirming the output of the functions being used are _accurate_. However they do _exist_ as a manual walk-through.

I agree, it would be nice if there were tests confirming output values. Maybe this could be addressed by providing a "demo script" that new users can run that automatically ingests data, corrects it, calculates some band depths, and visualizes. The documentation is great, but I suspect many users would appreciate having it all together in a script that can be run. That way it serves as both a form of test and a tutorial.
