### 2024-05-09 AStcherbinine/omegapy#9 @ryanbanderson
Tagging the JOSS thread: openjournals/joss-reviews#6566
