### 2024-05-15 979b357
add recursive search in subfolder for OMEGA files binaries
