+++ omegapy/omega_data.py
@@ -3,7 +3,7 @@
 
 ## omega_data.py
 ## Created by Aurélien STCHERBININE
-## Last modified by Aurélien STCHERBININE : 18/10/2023
+## Last modified by Aurélien STCHERBININE : 15/05/2024
 
 ##----------------------------------------------------------------------------------------
 """Importation and correction of OMEGA/MEx observations from binaries files.
@@ -259,7 +259,8 @@ class CubeError(Exception):
     def __init__(self, message):
         self.message = message
 
-def _readomega(cube_id, disp=True, corrV=True, corrL=True, data_path='_omega_bin_path'):
+def _readomega(cube_id, disp=True, corrV=True, corrL=True, data_path_qub='_omega_bin_path',
+               data_path_nav='_omega_bin_path'):
     """Python implementation of the `readomega.pro` routine from the SOFT10 OMEGA pipeline.
     
     Parameters
@@ -273,26 +274,32 @@ def _readomega(cube_id, disp=True, corrV=True, corrL=True, data_path='_omega_bin
         If `True`, compute the correction on the visible channel (Vis).
     corrL : bool, default True
         If `True`, compute the correction on the long-IR channel (L).
-    data_path : str, default _omega_bin_path
-        The path of the directory containing the data (.QUB) and 
-        navigation (.NAV) files.
+    data_path_qub : str, default _omega_bin_path
+        The path of the directory containing the data files (.QUB).
+    data_path_nav : str, default _omega_bin_path
+        The path of the directory containing the navigation files (.NAV).
 
     Returns
     -------
     out_data : dict
         {'ldat', 'jdat', 'wvl', 'ic', 'specmars'}
     out_geom : dict
-        {'latitude', 'longitude', 'emergence', 'incidence', 'altitude', 'ut_time',
-         'temperature, 'saturation_C', 'saturation_vis', 'lat_grid', 'lon_grid'}
+        {'latitude', 'longitude', 'emergence', 'emergence_norm', 'incidence',
+        'incidence_norm', 'altitude', 'ut_time', 'temperature_c', 'temperature_l',
+        'saturation_c', 'saturation_vis', 'lat_grid', 'lon_grid', 'latitude_v',
+        'longitude_v', 'altitude_v', 'incidence_norm_v', 'emergence_norm_v',
+        'lat_grid_v', 'lon_grid_v'}
     """
     # Default path
-    if data_path == "_omega_bin_path":
-        data_path = _omega_bin_path
+    if data_path_qub == "_omega_bin_path":
+        data_path_qub = _omega_bin_path
+    if data_path_nav == "_omega_bin_path":
+        data_path_nav = _omega_bin_path
     # Filename
     nomgeo0 = cube_id + '.NAV'
     nomfic0 = cube_id + '.QUB'
-    nomfic = os.path.join(data_path, nomfic0)
-    nomgeo = os.path.join(data_path, nomgeo0)
+    nomfic = os.path.join(data_path_qub, nomfic0)
+    nomgeo = os.path.join(data_path_nav, nomgeo0)
     if disp:
         print('\n\033[1mComputing OMEGA observation {0:s}\033[0m'.format(cube_id))
     # Orbit number in base 10
@@ -948,7 +955,7 @@ class OMEGAdata:
         self.add_infos = ''
 
         if not empty:
-            obs_name = uf.myglob(os.path.join(data_path, '*' + obs + '*.QUB'))
+            obs_name = uf.myglob(os.path.join(data_path, '**', '*' + obs + '*.QUB'), recursive=True)
             if obs_name is None:
                 print("\033[1;33mAborted\033[0m")
                 empty = True
@@ -1020,13 +1027,19 @@ class OMEGAdata:
[…]
