### 2024-05-15 32f5e91
add BD_omega function in omega_data
