### 2024-05-15 9ce72aa
fix BD_omega function in omega_data
