+++ omegapy/omega_data.py
@@ -3376,9 +3376,9 @@ def omega_mask(omega, emer_lim=None, inci_lim=None, tempc_lim=None, limsat_c=Non
 def BD_omega(omega, lam0, lamc1, lamc2, norm=True):
     """Compute the band depth on an OMEGA observation cube.
     Continuum linear between lamc1 and lamc2.
-
+    
     If an array is passed as argument for a wavelength value, the average is used.
-
+    
     Parameters
     ----------
     omega : OMEGAdata
@@ -3392,7 +3392,7 @@ def BD_omega(omega, lam0, lamc1, lamc2, norm=True):
     norm : bool, optional (default True)
         | True -> band_depth output is the normalized BD values.
         | False -> band_depth output is the BD values.
-
+    
     Returns
     -------
     band_depth : 2D array
