### 2024-05-15 6cffa4c
update doc to v3.0.5: recursive search and BD_omega
