### 2024-05-15 93dcc80
update docstring BD_omega
