### 2024-05-15 AStcherbinine/omegapy#9 @AStcherbinine
Thanks for the suggestion!
I've implemented an automatic recursive to handle potential subdirectories, also compatible with having the .NAV and .QUB files not in the same subfolder, as it can be found on the PSA with commit 979b3579b910f68d83429d1f676b5936fdc7fbb9
It has been deployed with version 3.0.5.
