### 2024-05-16 5dcf7b0
add statement of need in the documentation homepage
