### 2024-05-29 804136b
fix issue remaining lam_ma attribute
