### 2024-05-29 54ae1ca
add missing import numpy as np in doc example
