### 2024-05-29 f568fe9
add full example test script and update online doc
