### 2024-05-29 #6566 @AStcherbinine
> > The tests here are a bit weak in that they mainly revolve around confirming installation and turn users to the basic usage rather than confirming the output of the functions being used are _accurate_. However they do _exist_ as a manual walk-through.
>
> I agree, it would be nice if there were tests confirming output values. Maybe this could be addressed by providing a "demo script" that new users can run that automatically ingests data, corrects it, calculates some band depths, and visualizes. (Basically just taking the steps from the examples and putting them in a .py file) The documentation is great, but I suspect many users would appreciate having it all together in a script that can be run. That way it serves as both a form of test and a tutorial.

As suggested by reviewer @ryanbanderson, a full example script that can be run to [ load / apply atm & thermal corrections / compute band depth / and display the maps from the examples ] has been added as part of the test in commit [f568fe9](https://github.com/AStcherbinine/omegapy/commit/f568fe981fae8ede5e2b5b5f6a2778613792712b),
and the documentation has been updated accordingly: https://astcherbinine.github.io/omegapy/tests/#test-to-process-and-display-an-omega-observation

Test script file: [example_test_script.py](https://astcherbinine.github.io/omegapy/example_test_script.py)
