### 2024-06-19 60cb651
add workaround description for interactive plots in Google Colab Jupyter notebooks
