### 2024-07-10 0e67c8e
fix use_V_geom in interactive display
