### 2024-07-19 #6566 @AStcherbinine
Hello @dfm !
Thank you to you and the two reviewers for their time to this paper!

>     1. Take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially the author names and affiliations. I've taken a pass and it looks good to me!

Seems good to me!

>     2. Increment the version number of the software and report that version number back here.

Just moved the version number to 3.1.

>     3. Create an archived release of that version of the software (using Zenodo or something similar). _Please make sure that the metadata (title and author list) exactly match the paper._ Then report the DOI of the release back to this thread.

And here is the Zenodo DOI for the v3.1: https://doi.org/10.5281/zenodo.12775561
