### 2024-07-20 #6566 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
